Make HandlerLoader an interface
AbstractHandlerLoader to implement the interface as an abstract class.
Update the implementation handler loaders to reflect these changes.
